pyproject.toml
[tool.poetry]
name = "fhirguard"
version = "0.1.0"
description = ""
authors = ["Thomas Judd-Cooper <thomas.judd-cooper1@nhs.net>"]
readme = "README.md"
packages = [
    { include = "fhirguard_cli" },
    { include = "fhirguard_core" },
    { include = "fhirguard" }
]

[tool.poetry.scripts]
fhirguard = "fhirguard_cli.main:app"

[tool.poetry.dependencies]
python = ">=3.8"
typer = {extras = ["all"], version = "^0.9.0"}
requests = "^2.31.0"
fhir-resources = "^7.1.0"
platformdirs = "^4.1.0"
pycountry = "^23.12.11"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pytest = "^8.0.2"
ruff = "^0.2.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "C90", "I", "A", "C4", "T10", "PIE", "T20", "PYI", "PT", "RET", "SIM", "TID", "ARG", "PTH", "TD", "ERA", "PL", "PERF", "FURB"]
fixable = ["ALL"]
preview = true
ignore = ["B008"]

[tool.ruff.format]
docstring-code-format = true